final refactoring, client no longer launches game multiple times
