+++ electron/launcher/src/_actions/index.js
@@ -1,4 +1,3 @@
 // Central actions file
 export * from '_actions/user.actions';
 export * from '_actions/patcher.actions';
-export * from '_actions/client.actions';
