+++ electron/launcher/src/_actions/patcher.actions.js
@@ -2,7 +2,7 @@ import * as patcherTypes from '_constants/patcher.types';
 
 const ipc = window.require('electron').ipcRenderer;
 
-const checkForUpdate = (options) => {
+const checkForUpdate = (isInitialCheck) => {
   return (dispatch) => {
     dispatch({type: patcherTypes.CHECKING_FOR_UPDATE});
     ipc.send('fm-check-for-update');
@@ -10,11 +10,8 @@ const checkForUpdate = (options) => {
     ipc.on('fm-up-to-date', () => {
       dispatch({type: patcherTypes.IS_LATEST_VERSION});
 
-      if(options.initialCheck){
+      if(isInitialCheck){
         dispatch({type: patcherTypes.INITIAL_CHECK_COMPLETE});
-      }
-      if(options.preGameLaunchCheck){
-        dispatch({type: patcherTypes.QUEUE_GAME_LAUNCH});
       }
     });
 
@@ -35,13 +32,38 @@ const checkForUpdate = (options) => {
   }
 }
 
-const resetGameLaunchQueue = () => {
+const startGameClient = () => {
   return (dispatch) => {
-    dispatch({type: patcherTypes.RESET_GAME_LAUNCH_QUEUE});
+    dispatch({type: patcherTypes.CHECKING_FOR_UPDATE});
+    ipc.send('start-game-client');
+
+    ipc.on('fm-up-to-date', () => {
+      dispatch({type: patcherTypes.IS_LATEST_VERSION});
+      dispatch({type: patcherTypes.GAME_CLIENT_RUNNING});
+    });
+
+    ipc.on('fm-download-start', () => {
+      dispatch({type: patcherTypes.PATCHING})
+    });
+
+    ipc.on('fm-download-status-update', (e, update) => {
+      dispatch({
+        type: patcherTypes.DOWNLOAD_FILES_STATUS,
+        payload: update
+      });
+    });
+
+    ipc.on('start-game-client-success', () => {
+      dispatch({type: patcherTypes.GAME_CLIENT_EXIT});
+    });
+
+    ipc.on('start-game-client-fail', () => {
+      dispatch({type: patcherTypes.GAME_CLIENT_EXIT});
+    });
   }
 }
 
 export const patcherActions = {
   checkForUpdate,
-  resetGameLaunchQueue
+  startGameClient
 };
