--- electron/launcher/src/_constants/client.types.js
@@ -1,3 +1,0 @@
-export const START_GAME_CLIENT_SUCCESS = 'START_GAME_CLIENT_SUCCESS';
-export const START_GAME_CLIENT_FAIL = 'START_GAME_CLIENT_FAIL';
-export const CHECK_GAME_CLIENT_STATUS = 'CHECK_GAME_CLIENT_STATUS';
