+++ electron/launcher/src/_constants/patcher.types.js
@@ -1,7 +1,7 @@
 export const CHECKING_FOR_UPDATE = 'CHECKING_FOR_UPDATE';
 export const IS_LATEST_VERSION = 'IS_LATEST_VERSION';
 export const INITIAL_CHECK_COMPLETE = 'INITIAL_CHECK_COMPLETE';
-export const QUEUE_GAME_LAUNCH = 'QUEUE_GAME_LAUNCH';
 export const PATCHING = 'PATCHING';
 export const DOWNLOAD_FILES_STATUS = 'DOWNLOAD_FILES_STATUS';
-export const RESET_GAME_LAUNCH_QUEUE = 'RESET_GAME_LAUNCH_QUEUE';
+export const GAME_CLIENT_RUNNING = 'GAME_CLIENT_RUNNING';
+export const GAME_CLIENT_EXIT = 'GAME_CLIENT_EXIT';
